+++ src/dotnet/UI.Blazor.App/Components/ChatMessageEditor/IAttachmentWebFilePickerBackend.cs
@@ -2,5 +2,5 @@ namespace ActualChat.UI.Blazor.App.Components;
 
 public interface IAttachmentWebFilePickerBackend
 {
-    Task<bool> OnFilePicked(int id, string? fileName, string? fileType, int length);
+    Task<bool> OnFilePicked(int id, string? fileName, string? fileType, long length);
 }
